requirements.txt
Django
djangorestframework
drf-spectacular
psycopg2
django-cors-headers
python-binance
pandas
numpy
setuptools>=65.5.1
wheel>=0.38.1